requirements.txt
# =====================
# Core LLM + Orchestration
# =====================
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
langchain-groq
langchain-openai
langchain_chroma
langchain_tavily

docx2txt
python-multipart
pypdf

python-dotenv

pillow
pymupdf
faiss-cpu

openai>=1.40.0
tiktoken>=0.7.0

mcp>=1.0.0

faiss-cpu>=1.8.0
chromadb>=0.5.0
qdrant-client>=1.9.0

sentence-transformers>=3.0.0

sqlalchemy

requests>=2.32.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
tavily-python>=0.5.0

fastapi>=0.112.0
uvicorn>=0.30.0

pydantic>=2.7.0
numpy>=1.26.0
pandas>=2.2.0

aiohttp>=3.9.0
asyncio

loguru>=0.7.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0

tenacity>=8.3.0
python-dotenv>=1.0.0
orjson>=3.10.0

ragas>=0.1.0